Updating Kernel Defconfig Settings
1. Reversed the changes made to ma35d1_defconfig in
the commit dated September 5, 2023,
restoring it to its previous configuration.
2. Introduced ma35d1_random_base_defconfig,
which now supports the selection of the kernel's execution base address,
in addition to the default 0x80000.

If ma35d1_random_base_defconfig was utilized,
it is imperative to update all kernel modules (*.ko) accordingly.
This task will be carried out by either the Buildroot script or Yocto receipt.

Change-Id: Ia6ffd91f6a8c286fb9efc9687a87b646211bcecb