Update rpmsg and wormhole dirvers
In this BSP, Linux sends a message to the RTP M4 via the rpmsg driver,
which will invoke mailbox functions and, eventually, the wormhole
driver to send the message. The mailbox driver will queue a new
incoming message if the previous message was not completed.

In the event that the RTP M4 does not read the message from the wormhole,
the TX message will not complete. Consequently, all subsequent messages
cannot be sent and are queued in the mailbox queue, ultimately resulting
in a queue full condition.

To prevent the above situation, this patch adds a flush mechanism to
the MA35 rpmsg and wormhole drivers.

1. Update the wormhole driver
   - Add the 'wormhole_flush_data()' mailbox callback.

2. Update the rpmsg driver
   - Add 'mbox_flush()' to 'ma35d1_rpmsg_send()' and 'ma35d1_rpmsg_trysend()'
   to ensure that the message will be sent or timeout."

Change-Id: I4305082dc6b49037c56efdba42df0f13dd27b2b8