Resolved MA35 USB EHCI Host halt when entering suspend
When any USB device is connected to the USB Host port of the MA35 SoC,
the kernel cannot enter power down mode unless the device is unplugged.
We have identified and addressed the code causing this issue. It is a
workaround intended for some controllers, but it is not suitable for
the MA35 USB host controller. Removing this code resolves the issue.

Change-Id: I5a855a0fb5ca6addd2f5c1208338f86368607b40