Add USB PHY reset to the EHCI driver initialization function
The USB gadget driver executes USB PHY reset, but the USB host driver
does not. In the case where the USB gadget is disabled, the USB PHY
has not been reset. This can occasionally lead to USB host EHCI and
OHCI driver initialization failures, where the driver cannot complete
initialization, causing the USB host to malfunction.
This patch addresses this issue by performing a USB PHY reset in the
EHCI driver initialization function if the USB PHY is not yet ready.

Change-Id: Ic175b567b700718d4bf0916523d212f297eacebe